Handle the case of overflow or invalid parameters while tuning the step size.
I encountered an issue where, while tuning the step size, my model code was returning NaN. This caused Rcgmin's optimization to crash because it was assumed on line 289 that the value of f would be numeric (in order to compute the value of the condition in `if (f < min(fmin, f1))`). Replacing non-numeric values of f by Inf allows the step-size optimization to continue in this case.